remove data items in VinylDNSTestData; rebrand to helpers
