+++ modules/api/src/test/scala/vinyldns/api/backend/CommandHandlerSpec.scala
@@ -24,13 +24,15 @@ import org.mockito.Mockito._
 import org.mockito.{ArgumentCaptor, Mockito}
 import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{BeforeAndAfterEach, EitherValues, Matchers, WordSpec}
-import vinyldns.api.VinylDNSTestData
+import vinyldns.api.VinylDNSTestHelpers
 import vinyldns.api.backend.CommandHandler.{DeleteMessage, RetryMessage}
 import vinyldns.api.domain.dns.DnsConnection
 import vinyldns.core.domain.batch.BatchChangeRepository
 import vinyldns.core.domain.record.{RecordChangeRepository, RecordSetChange, RecordSetRepository}
 import vinyldns.core.domain.zone.{ZoneChange, ZoneChangeType, ZoneCommand, _}
 import vinyldns.core.queue.{CommandMessage, MessageCount, MessageId, MessageQueue}
+import vinyldns.core.TestRecordSetData._
+import vinyldns.core.TestZoneData._
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
@@ -40,7 +42,7 @@ class CommandHandlerSpec
     with Matchers
     with MockitoSugar
     with BeforeAndAfterEach
-    with VinylDNSTestData
+    with VinylDNSTestHelpers
     with EitherValues
     with EitherMatchers {
