+++ modules/api/src/test/scala/vinyldns/api/route/VinylDNSJsonProtocolSpec.scala
@@ -20,7 +20,7 @@ import org.joda.time.DateTime
 import org.json4s.JsonDSL._
 import org.json4s._
 import org.scalatest.{Matchers, WordSpec}
-import vinyldns.api.VinylDNSTestData
+import vinyldns.api.VinylDNSTestHelpers
 import vinyldns.core.domain.record._
 import vinyldns.core.domain.zone.{CreateZoneInput, UpdateZoneInput, ZoneConnection}
 
@@ -28,7 +28,7 @@ class VinylDNSJsonProtocolSpec
     extends WordSpec
     with Matchers
     with VinylDNSJsonProtocol
-    with VinylDNSTestData {
+    with VinylDNSTestHelpers {
 
   private val completeCreateZoneInput = CreateZoneInput(
     "testZone.",
